using size_t was just a bad idea, especially for x64 - revert to unsigned int

+++ ThrEd4/clip.h
@@ -27,38 +27,38 @@
 namespace clip {
 
 void    chnfn();
-void    clpbrd(size_t startVertex);
+void    clpbrd(unsigned int startVertex);
 void    clpic();
 void    clpout();
-void    delclps(size_t iForm);
-void    deleclp(size_t iForm) noexcept;
-void    delmclp(size_t iForm);
+void    delclps(unsigned int iForm);
+void    deleclp(unsigned int iForm) noexcept;
+void    delmclp(unsigned int iForm);
 void    duxclp();
-bool    isclp(size_t iForm) noexcept;
-bool    isclpx(size_t iForm) noexcept;
-bool    iseclp(size_t iForm) noexcept;
-bool    iseclpx(size_t iForm) noexcept;
-fPOINT* nueclp(size_t currentForm, size_t count) noexcept;
+bool    isclp(unsigned int iForm) noexcept;
+bool    isclpx(unsigned int iForm) noexcept;
+bool    iseclp(unsigned int iForm) noexcept;
+bool    iseclpx(unsigned int iForm) noexcept;
+fPOINT* nueclp(unsigned int currentForm, unsigned int count) noexcept;
 fPOINT* numclp() noexcept;
-void    oclp(const fPOINT* const clip, size_t clipEntries);
+void    oclp(const fPOINT* const clip, unsigned int clipEntries);
 
 namespace internal {
 
-	void clpcrnr(std::vector<fPOINT>& clipFillData, size_t vertex, const dPOINT& rotationCenter);
+	void clpcrnr(std::vector<fPOINT>& clipFillData, unsigned int vertex, const dPOINT& rotationCenter);
 	bool clpsid(const std::vector<fPOINT>& clipReversedData,
 	            std::vector<fPOINT>&       clipFillData,
-	            size_t                     start,
-	            size_t                     finish,
+	            unsigned int               start,
+	            unsigned int               finish,
 	            const dPOINT&              rotationCenter);
-	void clpsub(size_t fpnt, size_t cnt) noexcept;
+	void clpsub(unsigned int fpnt, unsigned int cnt) noexcept;
 	void clpxadj(std::vector<fPOINT>& tempClipPoints, std::vector<fPOINT>& chainEndPoints);
 	void duch(std::vector<fPOINT>& chainEndPoints);
-	void duchfn(const std::vector<fPOINT>& chainEndPoints, size_t start, size_t finish);
+	void duchfn(const std::vector<fPOINT>& chainEndPoints, unsigned int start, unsigned int finish);
 	void dufxlen(std::vector<fPOINT>& chainEndPoints);
 	void dulast(std::vector<fPOINT>& chainEndPoints);
 	void durev(std::vector<fPOINT>& clipReversedData);
 
-	size_t findclp(size_t formIndex) noexcept;
+	unsigned int findclp(unsigned int formIndex) noexcept;
 	void fxlen(std::vector<fPOINT>& chainEndPoints, const std::vector<double>& listSINEs, const std::vector<double>& listCOSINEs);
 	void fxlin(std::vector<fPOINT>&       chainEndPoints,
 	           const std::vector<double>& ListSINEs,
@@ -78,17 +78,21 @@ namespace internal {
 	             double                     clipAngle,
 	             dPOINT&                    moveToCoords,
 	             const dPOINT&              rotationCenter,
-	             size_t                     currentSide);
+	             unsigned int               currentSide);
 	void linsid(const std::vector<fPOINT>& clipReversedData,
 	            std::vector<fPOINT>&       clipFillData,
 	            double                     clipAngle,
 	            const dPOINT&              vector0,
 	            const dPOINT&              rotationCenter,
-	            size_t                     currentSide);
-	bool nupnt(double clipAngle, dPOINT& moveToCoords, size_t currentSide) noexcept;
-	void picfn(std::vector<fPOINT>& clipFillData, size_t start, size_t finish, double spacing, const dPOINT& rotationCenter);
+	            unsigned int               currentSide);
+	bool nupnt(double clipAngle, dPOINT& moveToCoords, unsigned int currentSide) noexcept;
+	void picfn(std::vector<fPOINT>& clipFillData,
+	           unsigned int         start,
+	           unsigned int         finish,
+	           double               spacing,
+	           const dPOINT&        rotationCenter);
 	bool ritclp(const std::vector<fPOINT>& clipFillData, const fPOINT& point);
-	void setvct(size_t start, size_t finish, double& ClipAngle, dPOINT& vector0) noexcept;
+	void setvct(unsigned int start, unsigned int finish, double& ClipAngle, dPOINT& vector0) noexcept;
 	void xclpfn(const std::vector<fPOINT>& tempClipPoints,
 	            const std::vector<fPOINT>& chainEndPoints,
 	            unsigned                   start,
