+++ ThrEd4/repair.h
@@ -14,8 +14,8 @@ void repar();
 
 namespace internal {
 
-	void     adbad(std::wstring& repairMessage, unsigned code, size_t count);
-	void     bcup(size_t find, BADCNTS& badData) noexcept;
+	void     adbad(std::wstring& repairMessage, unsigned code, unsigned int count);
+	void     bcup(unsigned int find, BADCNTS& badData) noexcept;
 	void     chkclp(const FRMHED* const formHeader, BADCNTS& badData);
 	void     chkeclp(const FRMHED* const formHeader, BADCNTS& badData);
 	void     chkfstch() noexcept;
